Split out registration and pw-reset tests methods
